+++ ytd/constants.py
@@ -1,0 +1,31 @@
+COMPAT_SOLL: list[str] = [
+            "sollumz_drawable",
+            "sollumz_fragment",
+            "sollumz_drawable_model",
+            "sollumz_drawable_dictionary",
+        ]
+
+ENV_TEXTURES: list[str] = [
+    "env_bark",
+    "env_cloth",
+    "env_crusty",
+    "env_noise_concrete",
+    "env_noise_heavy",
+    "env_smooth_concrete2",
+    "env_stucco",
+    "env_woodgrain",
+    "env_woodgrain_2",
+]
+
+SUPPORTED_FORMATS: list[str] = [
+    ".png",
+    ".jpg",
+    ".bmp",
+    ".tiff",
+    ".tif",
+    ".jpeg",
+    ".dds",
+    ".psd",
+    ".gif",
+    ".webp",
+]
