+++ ytd/misc.py
@@ -1,7 +1,6 @@
 from pathlib import Path
 import math
 
-SUPPORTED_FORMATS = [".png", ".jpg", ".bmp", ".tiff", ".tif", ".jpeg", ".psd", ".gif", ".webp"]
 
 
 def get_folder_list_from_dir(dir: str):
@@ -42,7 +41,7 @@ def closest_pow2_dims(width: int, height: int, max_dimension: int, make_half: bo
     use_height = height % max_dimension == 0
     
     if(use_width and use_height):
-        if( width / max_dimension < height / max_dimension):
+        if( width / max_dimension > height / max_dimension):
             num = width
         else:
             num = height
